# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ArtCafe.ai PubSub API
  description: |
    # ArtCafe.ai PubSub API

    This API provides access to the ArtCafe.ai PubSub service, which facilitates agent communication, management, and messaging.

    ## Authentication

    The API uses two authentication methods:
    
    1. **JWT Token Authentication**
       - Used for general API access
       - Pass the token in the `Authorization` header as a Bearer token: `Authorization: Bearer <your_token>`
    
    2. **SSH Key Authentication**
       - Used for agent authentication
       - Challenge-response based authentication flow
       - Provides JWT tokens for subsequent requests

    ## Multi-tenancy

    All resources are scoped to a tenant ID. The tenant ID must be provided:
    
    - In the `x-tenant-id` header or
    - In the JWT token payload

    ## Rate Limiting

    API rate limits are enforced based on the subscription tier of the tenant.
  version: 1.0.0
  contact:
    name: ArtCafe.ai Support
    email: support@artcafe.ai
    url: https://artcafe.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.artcafe.ai/v1
    description: Production server
  - url: https://api-dev.artcafe.ai/v1
    description: Development server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /agents:
    get:
      summary: List all agents
      description: Returns a list of all agents for the authenticated tenant
      tags:
        - Agents
      operationId: listAgents
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/nextTokenParam'
        - name: status
          in: query
          description: Filter by agent status
          required: false
          schema:
            type: string
            enum: [online, offline, idle, busy]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    
    post:
      summary: Register a new agent
      description: Registers a new agent for the authenticated tenant
      tags:
        - Agents
      operationId: createAgent
      security:
        - BearerAuth: []
        - tenantAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /agents/{agent_id}:
    get:
      summary: Get agent details
      description: Returns details for a specific agent
      tags:
        - Agents
      operationId: getAgent
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - name: agent_id
          in: path
          description: Agent ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /agents/{agent_id}/status:
    put:
      summary: Update agent status
      description: Updates the status of a specific agent
      tags:
        - Agents
      operationId: updateAgentStatus
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - name: agent_id
          in: path
          description: Agent ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentStatusUpdate'
      responses:
        '200':
          description: Agent status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /ssh-keys:
    get:
      summary: List all SSH keys
      description: Returns a list of all SSH keys for the authenticated tenant
      tags:
        - SSH Keys
      operationId: listSSHKeys
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/nextTokenParam'
        - name: agent_id
          in: query
          description: Filter by agent ID
          required: false
          schema:
            type: string
        - name: key_type
          in: query
          description: Filter by key type
          required: false
          schema:
            type: string
            enum: [access, agent, deployment]
      responses:
        '200':
          description: List of SSH keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHKeysResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    
    post:
      summary: Add a new SSH key
      description: Adds a new SSH key for the authenticated tenant
      tags:
        - SSH Keys
      operationId: createSSHKey
      security:
        - BearerAuth: []
        - tenantAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSHKeyCreate'
      responses:
        '201':
          description: SSH key successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /ssh-keys/{key_id}:
    get:
      summary: Get SSH key details
      description: Returns details for a specific SSH key
      tags:
        - SSH Keys
      operationId: getSSHKey
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - name: key_id
          in: path
          description: SSH key ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSH key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHKeyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    delete:
      summary: Delete SSH key
      description: Deletes a specific SSH key
      tags:
        - SSH Keys
      operationId: deleteSSHKey
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - name: key_id
          in: path
          description: SSH key ID
          required: true
          schema:
            type: string
      responses:
        '204':
          description: SSH key successfully deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /channels:
    get:
      summary: List all channels
      description: Returns a list of all channels for the authenticated tenant
      tags:
        - Channels
      operationId: listChannels
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/nextTokenParam'
        - name: agent_id
          in: query
          description: Filter by agent ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    
    post:
      summary: Create a new channel
      description: Creates a new channel for the authenticated tenant
      tags:
        - Channels
      operationId: createChannel
      security:
        - BearerAuth: []
        - tenantAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreate'
      responses:
        '201':
          description: Channel successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /channels/{channel_id}:
    get:
      summary: Get channel details
      description: Returns details for a specific channel
      tags:
        - Channels
      operationId: getChannel
      security:
        - BearerAuth: []
        - tenantAuth: []
      parameters:
        - name: channel_id
          in: path
          description: Channel ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  schemas:
...
    AgentCreate:
      type: object
      description: Agent creation model
      required:
        - name
      properties:
        name:
          type: string
          description: The agent name
        description:
          type: string
          description: The agent description
        metadata:
          type: object
          description: Additional metadata for the agent
      example:
        name: "My Agent"
        description: "A sample agent"
        metadata: {"version": "1.0", "capabilities": ["text", "image"]}
    AgentStatusUpdate:
      type: object
      description: Agent status update model
      required:
        - status
      properties:
        status:
          type: string
          enum: [online, offline, idle, busy]
          description: The new agent status
      example:
        status: "online"
    AgentResponse:
      type: object
      description: Agent response model
      required:
        - agent
        - success
      properties:
        agent:
          $ref: '#/components/schemas/Agent'
        success:
          type: boolean
          description: Whether the operation was successful
      example:
        agent:
          agent_id: "agent-1234567890"
          tenant_id: "tenant-1234567890"
          name: "My Agent"
          description: "A sample agent"
          status: "online"
          metadata: {"version": "1.0", "capabilities": ["text", "image"]}
          last_seen: "2023-07-01T12:00:00Z"
          created_at: "2023-06-01T12:00:00Z"
          updated_at: "2023-07-01T12:00:00Z"
        success: true
    AgentsResponse:
      type: object
      description: Agents list response model
      required:
        - agents
      properties:
        agents:
          type: array
          description: The list of agents
          items:
            $ref: '#/components/schemas/Agent'
        next_token:
          type: string
          description: The pagination token for the next page
      example:
        agents:
          - agent_id: "agent-1234567890"
            tenant_id: "tenant-1234567890"
            name: "My Agent"
            description: "A sample agent"
            status: "online"
            metadata: {"version": "1.0", "capabilities": ["text", "image"]}
            last_seen: "2023-07-01T12:00:00Z"
            created_at: "2023-06-01T12:00:00Z"
            updated_at: "2023-07-01T12:00:00Z"
        next_token: "next-page-token"
    
    # SSH key schemas
...
    SSHKeyCreate:
      type: object
      description: SSH key creation model
      required:
        - name
        - public_key
        - key_type
      properties:
        name:
          type: string
          description: The SSH key name
        public_key:
          type: string
          description: The public key in OpenSSH format
        key_type:
          type: string
          enum: [access, agent, deployment]
          description: The SSH key type
        agent_id:
          type: string
          description: The agent ID (required if key_type is 'agent')
        metadata:
          type: object
          description: Additional metadata for the SSH key
      example:
        name: "My SSH Key"
        public_key: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC..."
        key_type: "agent"
        agent_id: "agent-1234567890"
        metadata: {"purpose": "agent authentication"}
...
    SSHKeyResponse:
      type: object
      description: SSH key response model
      required:
        - key
        - success
      properties:
        key:
          $ref: '#/components/schemas/SSHKey'
        success:
          type: boolean
          description: Whether the operation was successful
      example:
        key:
          key_id: "key-1234567890"
          tenant_id: "tenant-1234567890"
          name: "My SSH Key"
          public_key: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC..."
          fingerprint: "SHA256:abcdefghijklmnopqrstuvwxyz123456789"
          key_type: "agent"
          agent_id: "agent-1234567890"
          status: "active"
          last_used: "2023-07-01T12:00:00Z"
          revoked: false
          created_at: "2023-06-01T12:00:00Z"
          updated_at: "2023-07-01T12:00:00Z"
        success: true
    SSHKeysResponse:
      type: object
      description: SSH keys list response model
      required:
        - keys
      properties:
        keys:
          type: array
          description: The list of SSH keys
          items:
            $ref: '#/components/schemas/SSHKey'
        next_token:
          type: string
          description: The pagination token for the next page
      example:
        keys:
          - key_id: "key-1234567890"
            tenant_id: "tenant-1234567890"
            name: "My SSH Key"
            public_key: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC..."
            fingerprint: "SHA256:abcdefghijklmnopqrstuvwxyz123456789"
            key_type: "agent"
            agent_id: "agent-1234567890"
            status: "active"
            last_used: "2023-07-01T12:00:00Z"
            revoked: false
            created_at: "2023-06-01T12:00:00Z"
            updated_at: "2023-07-01T12:00:00Z"
        next_token: "next-page-token"
    
    # Channel schemas
...
    ChannelCreate:
      type: object
      description: Channel creation model
      required:
        - name
      properties:
        name:
          type: string
          description: The channel name
        description:
          type: string
          description: The channel description
        metadata:
          type: object
          description: Additional metadata for the channel
      example:
        name: "General"
        description: "General discussion channel"
        metadata: {"type": "general", "public": true}
    ChannelResponse:
      type: object
      description: Channel response model
      required:
        - channel
        - success
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        success:
          type: boolean
          description: Whether the operation was successful
      example:
        channel:
          channel_id: "channel-1234567890"
          tenant_id: "tenant-1234567890"
          name: "General"
          description: "General discussion channel"
          metadata: {"type": "general", "public": true}
          created_at: "2023-06-01T12:00:00Z"
          updated_at: "2023-07-01T12:00:00Z"
        success: true
    ChannelsResponse:
      type: object
      description: Channels list response model
      required:
        - channels
      properties:
        channels:
          type: array
          description: The list of channels
          items:
            $ref: '#/components/schemas/Channel'
        next_token:
          type: string
          description: The pagination token for the next page
      example:
        channels:
          - channel_id: "channel-1234567890"
            tenant_id: "tenant-1234567890"
            name: "General"
            description: "General discussion channel"
            metadata: {"type": "general", "public": true}
            created_at: "2023-06-01T12:00:00Z"
            updated_at: "2023-07-01T12:00:00Z"
        next_token: "next-page-token"
    
    # Tenant schemas
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        path:
          type: string
          description: Path to the erroneous field
      example:
        detail: "Invalid request"
        code: "BAD_REQUEST"
        path: "name"
  parameters:
    limitParam:
      name: limit
      in: query
      description: Maximum number of results to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    nextTokenParam:
      name: next_token
      in: query
      description: Pagination token for the next page
      required: false
      schema:
        type: string
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid request"
            code: "BAD_REQUEST"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Authentication required"
            code: "UNAUTHORIZED"
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Access denied"
            code: "FORBIDDEN"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Resource not found"
            code: "NOT_FOUND"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    tenantAuth:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant ID header